+++ src/browser/js/app/widgets/push.js
@@ -59,19 +59,19 @@ module.exports.create = function(widgetData,container) {
 
     widget.getValue = function() {
         return widget[widget.lastChanged] ?
-            widgetData.on.value !== undefined ? widgetData.on.value : widgetData.on
+            widgetData.on != null && widgetData.on.value !== undefined ? widgetData.on.value : widgetData.on
             :
-            widgetData.off.value !== undefined ? widgetData.off.value : widgetData.off
+            widgetData.off != null && widgetData.off.value !== undefined ? widgetData.off.value : widgetData.off
     }
 
     widget.setValuePrivate = function(v,options={}) {
-        if (v===widgetData.on || (v.value === widgetData.on.value && v.value !== undefined)) {
+        if (v===widgetData.on || (widgetData.on != null && v.value === widgetData.on.value && v.value !== undefined)) {
             widget.addClass('active')
             widget.active = 1
             if (options.send) widget.sendValue(v)
             widget.lastChanged = 'active'
             if (options.sync) widget.trigger({type:'sync',id:widgetData.id,widget:widget, linkId:widgetData.linkId, options:options})
-        } else if (v===widgetData.off || (v.value === widgetData.off.value && v.value !== undefined)) {
+        } else if (v===widgetData.off || (widgetData.off != null && v.value === widgetData.off.value && v.value !== undefined)) {
             widget.removeClass('active')
             widget.active = 0
             if (options.send) widget.sendValue(v, widgetData.norelease)
@@ -85,13 +85,13 @@ module.exports.create = function(widgetData,container) {
             if (options.send || options.sync) widget.setValuePrivate(v,options)
             return
         }
-        if (v===widgetData.on || (v.value === widgetData.on.value && v.value !== undefined)) {
+        if (v===widgetData.on || (widgetData.on != null && v.value === widgetData.on.value && v.value !== undefined)) {
             widget.addClass('on')
             widget.state = 1
             if (options.send) widget.sendValue(v)
             widget.lastChanged = 'state'
             if (options.sync) widget.trigger({type:'sync',id:widgetData.id,widget:widget, linkId:widgetData.linkId,options:options})
-        } else if (v===widgetData.off || (v.value === widgetData.off.value && v.value !== undefined)) {
+        } else if (v===widgetData.off || (widgetData.off != null && v.value === widgetData.off.value && v.value !== undefined)) {
             widget.removeClass('on')
             widget.state = 0
             if (options.send) widget.sendValue(v)